refactor(optimizer): split Boring tail trimming region into RecordingOptimizer.TailTrim.cs
